[FIX] helpdesk: sort tickets by ID by default in portal
In the ticket list view of the portal, we can filter tickets by create date and by ID.
This is a bit redundant, and sorting by id reflects more accurately what we want to see.
Therefore, this commit removes the filter by create date to default to the filter by ID instead.

Task-4438387

closes odoo/enterprise#77577

Related: odoo/odoo#194695
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>