[FIX] knowledge: fix article commands tour
1st issue:
Since the introduction of the [media command] in the editor, replacing the
previous `/file`, some tour steps of a disabled test became obsolete.

This commit fixes these steps according to changes introduced along the `/media`
command, in order to re-enable this test.

2nd issue:
Other steps were incorrectly modified by [this commit] and had to be adapted
too, related to the composer signature.

3rd issue:
An embedded video was present in an article body in the readonly tour, and the
patch supposed to neutralize it (to avoid external http requests) was applied in
a manner that resulted in an indeterministic result: Sometimes the article body
was loaded with the video, resulting in errors, sometimes the patch was applied
resulting in the tour performing normally. This commit fixes the issue  by
ensuring there is no embedded video after the end of the editable tour, so that
there is no need to care about it in the readonly tour.

[media command]: https://github.com/odoo/odoo/commit/96c8c398c0fbef519b56edf4941691d11372eac0
[this commit]: https://github.com/odoo/enterprise/commit/1a73e227a5ce9fb1194e5664c8c4f5fcfe51b572

runbot-task-111948

closes odoo/enterprise#77769

X-original-commit: 3a1c846abde46891fcbc5c4c6828f38e28d2677d
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>