[REF] pos_preparation_display: refactoring of module
- The pos_preparation_display has been moved to pos_enterprise since
it is installed if point_of_sale is installed in an enterprise environement.

- The preparation order_stage model is rename in pos_prep_state and is now
specific to an preparation orderline and not order anymore. The done field
is now infer by checking if the state is in the last stage and todo false.
The todo field has been moved to the state model since it is different per
orderline per preparation display.

- Related models is now introduced in the preparation display.

- The logic in the preparation display front end is mainly done using the
states of the preparation orderlines.

- pos_hr_preparation_display has been removed since it's not used anymore.

closes odoo/enterprise#78493

Related: odoo/odoo#201170
Related: odoo/upgrade#7499
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>