[FIX] hr_payroll_account: prevent payslip stage reset on create entry
In the tree view of the payslip model, we can do mass edit to create a journal
entry. Before, we were able to perform this action on a paid payslip which does
not make sense. Now we have an error message if we try to do it.

closes odoo/enterprise#77677

Task: 3874129
Related: odoo/upgrade#7167
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>