[IMP] sale_commission: ajustments linked to plan achievements
Before this commit, the adjustments used the type field to determine
which plan achievement could be used. Unfortunately, it would trigger
issues when several plan achievements would match. For Example:
Adjustment with amount invoiced
The related plan will be checked based on company and approved state.
The plan achievement lines will match the type but if the plan has these
lines:

- amount_invoiced, product foo rate 25%
- amount_invoiced, product category bar 50%

As a result the adjustment would be used twice in the achievepment
report and the commission would be too high.

This commit force the adjustment to be linked to a single achievement
line of a plan. Therefore the adjustment can only be used once.

Taskid: 4471762
X-original-commit: 20b0a539e484a3f5e32b8a418939c803980d1522
Part-of: odoo/enterprise#79887
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>