[IMP] documents: allow owners to delete any document
It's weird to prevent deletion of one's own document
because they are not editor of the folder. This rule
should only apply to editor members, not owners.

We're also fixing the `unlink` flow which allowed users
to delete documents in folders where they had 'read'
permission contrary to what we had implemented in
`action_archive`. We're now aligning this logic in
`unlink` for active documents.
Note that we consider OK to allow a user A to delete a
folder in the trash if it contains document that they
don't have access to, as it must have been put there
by someone who could in the first place, and the cron
would eventually delete everything automatically.

Task-4510838

X-original-commit: 0c46b6d05c342a5edb6f06a7e5868b3ab73827b0
Part-of: odoo/enterprise#79282
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>