[IMP] marketing_automation: x2m field kanban: replace `Add` button with a Kanban card
This commit, aims change the look of the `Add` button. From now, it's a
kanban card at the end of the Kanban. The label is "Add" followed by the
model name of the field.

The `Add` card appears only when the field match these condition:
* not readonly
* the kanban has the attribute "create=true" (default value)
* the kanban has the attribute "link=true" (default value)
* the kanban has no "control"

Note we have also adapt some tests and some views due to this change

task-4361655

closes odoo/enterprise#77746

Related: odoo/odoo#194470
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>