[IMP] documents: reintroduce translatable name and sequence
Make documents name translatable
================================

Before Documents' makeover (18.0), it was possible to translate
folder names.

This commit makes the name field of 'documents.document'
translatable. As folders and documents share the same model, document
names will also be translatable.

Reorder folders in search panel
===============================

This commit reintroduces the possibility to reorder folders in the
search panel using drag & drop:
- Document manager can reorder folders inside "Company".
- Internal users can reorder folders in "My Drive".

A sequence field (integer) is added to the documents model to implement
the feature.

The feature works only when you drag a folder from and drop to the
search panel.

Task-4512349

closes odoo/enterprise#77890

Related: odoo/upgrade#7133
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>
Co-authored-by: Florian Charlier <flch@odoo.com>