[FIX] l10n_ke_edi_oscu: handle missing product name to avoid zerodivision error
This error occurs when clicking the "Send to eTIMS" button because the move
lines lack a product name.

Steps to reproduce:
- Install 'l10n_ke_edi_oscu' module > Switch to 'KE Company'
- Create a NEW Vendors Bill > Add a Product and remove the product name > Save
- Click 'Send to eTIMS'

Traceback:
ZeroDivisionError: float division by zero

At [1], we encounter 'product = False' because, at [2], the product_id is being
retrieved from the move lines, but no product name is present in the move lines.

After this commit, we will ensure that valid values are present in product lines;
otherwise, an warning will be raised.

[1] - https://github.com/odoo/enterprise/blob/4832c81b89f8831d5a06671ee9c06ac06973b7b5/l10n_ke_edi_oscu/models/account_move.py#L265
[2]- https://github.com/odoo/enterprise/blob/4832c81b89f8831d5a06671ee9c06ac06973b7b5/l10n_ke_edi_oscu/models/account_move.py#L240

sentry-6242406302

closes odoo/enterprise#78093

X-original-commit: eed016bb80fa1fba71ec50ce332e9237c6bb5804
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Aayush Modi (aamo) <aamo@odoo.com>