[FIX] account_online_synchronization: Find duplicate tool traceback
Fix traceback when changing the date in the find duplicate tool

When the user changes the date in the find duplicate tool, a
traceback occurs when transitioning from no rows displayed to
rows being displayed.

This issue is caused by the `getRowClass` function referencing
the outdated `firstIdsInGroup` field, which is never updated.
If there were no values initially, the function attempts to call
`includes` on a `false` value, resulting in the error.

This commit resolves the issue by removing the obsolete
`firstIdsInGroup` field and adding a condition to ensure that
the value being checked is an array, preventing any traceback.

no task id

closes odoo/enterprise#77925

X-original-commit: 24b2a4fbf069f3d11eb72a5df6a727bdeb17b08a
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>