[FIX] web_studio: reportEditor: better rendering in preview mode
Before this commit the footer of an external layout was sometimes overlaying the contents of the report
This was due to the fact that the footer was in set to position=fixed (to make it sticky to the bottom of the iframe)

This commit changes the strategy. It makes, just within the report editor, the footer's parent in display flex to be
able to apply an auto margin on the footer.

After this commit, we trade the accuracy of the iframe's size representing paper size to the UX of editing a report
on which many t-ifs nodes are visible and take space, even though they won't all be visible at actual render time.

opw-4520367

closes odoo/enterprise#80402

X-original-commit: 8efd63af9ec85ba126fa029ced9795b44a7713fb
Related: odoo/odoo#199656
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>