[FIX] account: outbound shouldnt require validated bank account
The require_partner_bank_account flag on the payment shouldn't require
the allow_out_payment flag on it when the payment type is inbound as it
leads to confusion.

This generates an issue with the l10n_it_riba RiBa payment method. It
requires the bank account, but the money is incoming, not outgoing.

task-4497749

closes odoo/enterprise#79536

X-original-commit: c2d5703879b342623adc8b7cd51eadce3d6f4fb2
Related: odoo/odoo#198048
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Paolo Gatti (pgi) <pgi@odoo.com>