[IMP] l10n_br_avatax{,_sale}: override of purpose of use
In some cases the purpose of use will change based on the exact
conditions of a sale (e.g. the partner, ...). An example is if you do
b2b and b2c sales, the b2b sale's purpose will be "resale" and the b2c
sale's purpose could be "use or consumption". To avoid users having to
duplicate their products we provide an optional override on sale.order
and account.move.

task-4520757

closes odoo/enterprise#78050

Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>