[FIX] web_studio: fix test failing in all apps
Before this commit, when all apps were installed, the tour executed a step that triggered a rpc and a re-render.
But the next step had its trigger set to a node already in the dom before the re-render.

So, at the step after, the dropdown which was clicked just before had disappeared.

By filtering even more and by applying more precise selectors, we fix this issue

runbot-error-108304

closes odoo/enterprise#78208

X-original-commit: 9f750bb32d9252a735a9ef39af5192346b644af3
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>