[FIX] documents(,_account): allow un-embedding actions
Reproduce:
1. Install documents_account
2. Go to the finance folder
3. Try to un-embedd the action "Create a Vendor Bill"
4. You can't

This occurs because we created the embedded actions by passing a
group in the data, but we can't do that in the UI, so we cannot
un-embed the existing action.

This also means that groups cannot be enforced at the embedded
actions level using their groups_id field, but rather the groups_id
field of the related server action, and it was actually redundant.
Thankfully these records are correct.

Task-4525850

X-original-commit: 4be7ee634dd52351816ca786a1882bdfead97ff3
Part-of: odoo/enterprise#81007
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>