[FIX] documents: fix list and kanban range selection
Before this commit, In kanban view when selecting multiple documents while
holding shift as modifier, it could happen that folders were selected too.
This was due to the logic which was using document ids as the range value when
looping on `root.records` but with the new UI where the folders are now
displayed on top of the documents it was buggy.

In list view, we would like to keep the same behavior as in other list views.

This commit fixes those issues.

Task-4532046

closes odoo/enterprise#78861

X-original-commit: fa81139695c3851f2f6c334e5152fd57977f3658
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>