[FIX] l10n_de_intrastat: fix test
The intrastat xml for germany should include the supplementary_unit_code value, which is already present in the xml template, but wasn't returned by the intrastat query. As the field is added here: https://github.com/odoo/enterprise/pull/71114, we can now fix the expected xml in the test.

task-4194404

closes odoo/enterprise#71114

closes odoo/enterprise#81090

X-original-commit: 9a450a5a846af5353c73c7e3b1460eb8ad843c67
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Igor Bertrand (igbe) <igbe@odoo.com>