[FIX] account_accountant: Fix missing exchange diff with CABA on bank rec widget
The 'no_exchange_difference' is used to prevent exchange diff in the whole reconciliation.
However, the reconciliation is recursive when dealing with CABA entries.
From the bank reconciliation widget, we want to prevent the exchange diff only for the top
reconciliation but not recursively. Otherwise, the CABA entry ends up with a residual amount
but considered as fully reconciled.

Instead, let's use 'no_exchange_difference_no_recursive'.

See the test in this commit.

closes odoo/enterprise#79138

Task-id: 4531511
X-original-commit: 392d3989c4a5219c90eacff751ce1dfdfb44df41
Related: odoo/odoo#197408
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>