[FIX] documents: fix control panel wrong display
When you reduce the screen width or view the page on devices with
medium/small screens, the control panel button group rearranges itself
in an unsightly way (e.g. excessive button height).

It also looks like the Documents mobile view is too different from
the standard Odoo mobile view.

Medium breakpoint and below:
===========================

All actions of the selected documents are grouped together in a single
dropdown (in the control panel).

Small devices, portrait:
=====================================

We make the look and feel closer to the standard Odoo list view for
mobile, but we do it also for the kanban view as, in Documents, kanban
cards are selectable:
- The action dropdown for the selected documents takes the place of the
DocumentCogItemMenu.
- The list selection box floats to the top left of the navbar.
- The details panel toggler in the control panel navigation disappears.

task-4471881

closes odoo/enterprise#78210

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>