[IMP] account_reports: updated PnL report traceback
The aim of this commit is to avoid having inconsistent data leading to a
Traceback while loading the lines of the Profit and Loss report.

Context:
cbe74884b937f630cf1d2ccf475a04e70b6f4669 changed the structure hierachy.
It wasn't a big issue until e96e8d4aa663a4a5305214f2124243070fa797bb
resequenced the lines and changed the order of computation in
`<account.report>._get_line`. This method assumes the parent lines are
going to be computed first and then cached in a dict.
Leaving `parent_id` made this assumption wrong leading to a KeyError
exception throwing a traceback.

Before the commit:
The data kept the previous `parent_id` which could lead to a traceback
due to inconsistent data.

After the commit:
`parent_id` not being set, the code doesn't try to get it from the
cached lines and computes it correctly.

closes odoo/enterprise#78336

Task-id: None
X-original-commit: 9caaed57e34d5c931015f1614c63ab537df853ab
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>