[FIX] l10n_be_hr_payroll: fix traceback for sales on payroll
Bug:
When creating a new payslip for a salesperson with work entries of code
LEAVE1731, we would get a traceback.

Reason:
Since we were creating the record and did not yet save, the code was calling a
method that did an SQL query on ids of self which is empty.

Fix:
We created a new method that uses the orm instead to handle this specific case.

closes odoo/enterprise#82725

X-original-commit: 04df4189ba53c0b692de8c83c32c9366f81c97fe
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>