[FIX] hr_payroll : remove search_default_partner_id from act_contribution_reg_payslip_lines
This argument causes several issues and do not make sense as there is already a default_partner_id.

For example, you click on payslip lines from the contact Abigail Peterson, and it filters everything
with a search on partner_id = 1, which doesn't make sense.

It also causes issues if the partner with id 1 doesn't exists.

opw-4516239

closes odoo/enterprise#86569

X-original-commit: eb6cd7fd85c2e3a0e7681ecd700e0746e586b9e9
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Loïc Leloup (lole) <lole@odoo.com>