[IMP] sale_amazon: sell products in FBA and FBM
Amazon allows you to sell products both in FBA and in FBM (as a fallback
when Amazon does not have stock anymore), but currently this
configuration is not working with Odoo because when an offer is
considered FBA, then the FBM stock is forced to 0 in Amazon.

This commit removes this limitation, allowing the stock to be correctly
synchronized with Amazon in all cases (FBA, FBM, both FBA and FBM).

task-4092410

closes odoo/enterprise#78293

Related: odoo/upgrade#7507
Signed-off-by: Morgane Demesmaeker <edm@odoo.com>