[IMP] documents_account: Remove query count
Here a change from a M2o to a O2m relationship in `hr_expense` between
the models `account.move` and `hr.expense(.sheet)`
would trigger this query count, as unlinking a move would at least add
a SELECT query to search for expense referencing the move.

Because this would arise every time a O2m or a query would be added
on an account.move unlink this would break the test a lot for changes
outside of this test scope. Localizations tasks for example would
break it quite often.

Part-of: odoo/enterprise#75507
Related: odoo/odoo#189701
Related: odoo/upgrade#6990
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>